Allow "globbing" filenames
This is useful for a case where the filename is not known, such as when
a migration is generated in Rails and we want to write into the
migration file. We can replace the migration number with "*" to glob the
filename to append/write/remove.

If there are more than one file matching the glob, an error is raised.