Update syntax
The previous output directives (i.e. `:::=`) were similar to ERB, but did not allow a way to render only the result and not the command. To allow for this syntax I've adopted the use of the characters `>` (arrow) and `-` (dash).

These indicate, render (`>`) and do-not render (`-`) respectively. Their positions correspond to code command and result so `:::->` indicates you do not want to command to render, but do want the result.

Here are our four different cases:


- `:::>`  (yes command, not result)
- `:::>>` (yes command, yes result)
- `:::-`  (not command, not result)
- `:::->` (not command, yes result)