Allow running tasks in the Background
Sometimes you want to start a long lived process like a server in the background. In that case, the `background` namespace has your, well, back.

To start a process, pass in the command as the first arg, and give it a name (so it can be referenced later):

```
:::>> background.start("rails server", name: "server")
```

You can make the background process wait until it receives a certain string in the logs. For instance to make sure that the server is fully booted:

```
:::>> background.start("rails server", name: "server", wait: "Listening on")
```

You can stop the process by referencing the name:

```
:::-- background.stop(name: "server")
```

You can also get the log contents:


```
:::>> background.log.read(name: "server")
```

You can also truncate the logs:

```
:::>> background.log.clear(name: "server")
```