Fix Log deletion bug
A `Tempfile` works by defining a "finalizer" this is code that gets run when the object is cleaned up via GC. Here's a link:

https://github.com/ruby/ruby/blob/935d0b3d05dfc8b30bba505792129bf0e33ebe3b/lib/tempfile.rb#L132

As long as another object has a reference to a Tempfile then it will not be deleted from disk. In this scenario we're passing it to a `Pathname` object and it does not retain a reference to the passed in argument. You can reproduce this bug/behavior outside of Rundoc:

```
require 'pathname'
require 'tempfile'

def make_pathname(value = http://Tempfile.new("log"))
  http://Pathname.new(value)
end

p = make_pathname
FileUtils.touch(p)

puts "File: #{p.file?}"

GC.start

puts "File: #{p.file?}"
```

This generates:

```
$ ruby scratch.rb
File: true
File: false
```

This PR fixes the behavior by forcing a reference to the Tempfile object in the top level ProcessSpawn instance instead of expecting the Pathname `@log` instance variable to hold its reference.