Allow pipe commands to start with $
Pipe commands can take two forms:

```
| tail
```

and:

```
| $ tail
```

So that fully qualified commands can be used. There was a bug that commands that started with $ were not being used.

Prior to this the parser was expecting a `command` which looks like this:

```
    # :::>> $ cat foo.rb
    rule(:command) {
      (
        start_command >>
        visability.as(:cmd_visability) >> spaces? >>
        method_call.as(:cmd_method_call) >> newline.maybe #>> match(/\z/)
      ).as(:command)
    }
```

Which would never be matched. So we were accidentally relying on the failure rescue mode from the parser to give us the behavior for `| tail`.

With this change any commands should be allowed such as `| $ tail`.