Refactor env[:before] to hold an array
The env[:before] interface allows commands to annotate outside of the code block. Previously this was a string that also contained the code fence information. This work splits the information into a new keys `:fence_start` and `:fence_end` and allows information to be appended directly to env[:before] instead of having to manually insert the prior before value.