Allow code commands to modify stack
There's many different places that can execute code now, so it's a bit confusing when trying to add new features or do things like partial results are written to disk in the event of a failure (this is still basically broken due to this problem). Instead of having commands that actually execute other commands instead, they can now modify the stack by pushing another command onto it and that will be executed next in line.